Reuse filter selection view from Filters modal (for the Segments modal)
