--- web/app/pages/Project/tabs/Traffic/TrafficHeaderActions.tsx
@@ -1,13 +1,0 @@
-import ProjectViewHeaderActions, {
-  ExportType,
-} from '../../View/components/ProjectViewHeaderActions'
-
-interface TrafficHeaderActionsProps {
-  exportTypes: ExportType[]
-}
-
-const TrafficHeaderActions = ({ exportTypes }: TrafficHeaderActionsProps) => {
-  return <ProjectViewHeaderActions tnMapping={{}} exportTypes={exportTypes} />
-}
-
-export default TrafficHeaderActions
